+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - decoding of capability status
 - new charge modes
 - new plug states
+- new engine and car types
 
 ## [0.27.2] - 2021-12-16
 - No changes, just a new version number for pypi